Refactor parser complete
