+++ sources/parser/parser_operations.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   parse_operations.c                                 :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: qbeukelm <qbeukelm@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/01/17 15:54:28 by quentinbeuk       #+#    #+#             */
-/*   Updated: 2024/01/19 16:59:16 by qbeukelm         ###   ########.fr       */
+/*                                                        ::::::::            */
+/*   parser_operations.c                                :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: quentinbeukelman <quentinbeukelman@stud      +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2024/01/25 22:42:15 by quentinbeuk   #+#    #+#                 */
+/*   Updated: 2024/01/25 22:42:21 by quentinbeuk   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
