+++ tests/test_commands.txt
@@ -1,0 +1,13 @@
+
+=== ONLY PIPES ===
+
+test this | and that | three more
+one 1 2 3 4 5 | two 1 | three 1 | four 1 2 3 | five 1 | six 1 | seven 1
+one 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 | two 1 2 3 | three 1
+
+
+=== PIPES AND REDIRECTS ===
+
+one 1 2 > outfile | two 1 2 3 | three 1 2
+one 1 2 > outfile1 > outfile2 > outfile3  3 4 5 6 | two 1 2 3 | three 1 2
+
